+++ src/Instances.js
@@ -5,7 +5,6 @@ import {
   keyBy,
   get,
   template,
-  isEmpty,
 } from 'lodash';
 import {
   injectIntl,
@@ -68,14 +67,14 @@ const filterConfig = [
 ];
 
 const searchableIndexes = [
-  { label: 'ui-inventory.search.all', value: 'all', queryTemplate: 'title="%{query.query}" or contributors =/@value "%{query.query}" or identifiers =/@value "%{query.query}"' },
+  { label: 'ui-inventory.search.all', value: 'all', queryTemplate: 'title="%{query.query}" or contributors =/@name "%{query.query}" or identifiers =/@value "%{query.query}"' },
   { label: 'ui-inventory.barcode', value: 'item.barcode', queryTemplate: 'item.barcode=="%{query.query}"' },
   { label: 'ui-inventory.instanceId', value: 'id', queryTemplate: 'id="%{query.query}"' },
   { label: 'ui-inventory.title', value: 'title', queryTemplate: 'title="%{query.query}"' },
   { label: 'ui-inventory.identifier', value: 'identifier', queryTemplate: 'identifiers =/@value "%{query.query}"' },
-  { label: 'ui-inventory.isbn', prefix: '- ', value: 'isbn', queryTemplate: 'identifiers =/@identifierTypeId="<%= identifierTypeId %>" "%{query.query}"' },
-  { label: 'ui-inventory.issn', prefix: '- ', value: 'issn', queryTemplate: 'identifiers =/@identifierTypeId="<%= identifierTypeId %>" "%{query.query}"' },
-  { label: 'ui-inventory.contributor', value: 'contributor', queryTemplate: 'contributors =/@value "%{query.query}' },
+  { label: 'ui-inventory.isbn', prefix: '- ', value: 'isbn', queryTemplate: 'identifiers =/@value/@identifierTypeId="<%= identifierTypeId %>" "%{query.query}"' },
+  { label: 'ui-inventory.issn', prefix: '- ', value: 'issn', queryTemplate: 'identifiers =/@value/@identifierTypeId="<%= identifierTypeId %>" "%{query.query}"' },
+  { label: 'ui-inventory.contributor', value: 'contributor', queryTemplate: 'contributors =/@name "%{query.query}"' },
   { label: 'ui-inventory.subject', value: 'subject', queryTemplate: 'subjects="%{query.query}"' },
 ];
 
@@ -281,18 +280,6 @@ class Instances extends React.Component {
     this.props.updateLocation({ qindex });
   }
 
-  onFilterChange = (e) => {
-    console.log(e)
-    //
-    // // if we're clearing the filters, clear the query index too
-    if (isEmpty(e)) {
-      this.props.updateLocation({ qindex: 'all' });
-    }
-
-    // const qindex = e.target.value;
-    // this.props.updateLocation({ qindex });
-  }
-
   updateFilters(prevFilters) { // provided for onChangeFilter
     const filters = Object.keys(prevFilters).filter(key => filters[key]).join(',');
     this.props.updateLocation({ filters });
@@ -421,7 +408,6 @@ class Instances extends React.Component {
           selectedIndex={get(this.props.resources.query, 'qindex')}
           searchableIndexesPlaceholder={null}
           onChangeIndex={this.onChangeIndex}
-          filterChangeCallback={this.onFilterChange}
           filterConfig={filterConfig}
           initialResultCount={INITIAL_RESULT_COUNT}
           resultCountIncrement={RESULT_COUNT_INCREMENT}